Refactor: more details in comments on this commit (userRegister.html): change name (emailDelivery.js): move function in their file change compare undifine instead of null when check record finds in detabase or not
