+++ client/src/views/userRegister.html
@@ -44,7 +44,7 @@
                 <div class="form-section" id="resetpass-section">
                     <p>Enter your email to receive the password recovery link</p>
                     <input type="email" name="email" id="email" placeholder="Email">
-                    <button id="submit-btn" class="primary-btn">Send</button>
+                    <button id="resetpass-btn" class="primary-btn">Send</button>
                 </div>
             </div>
         </div>
