+++ server/routes/authRouters.js
@@ -4,7 +4,8 @@
 
 import {Router} from 'express';
 import {ReqSignUpDto, ReqSignInDto} from '../src/dtos/userRegisterDto.js';
-import {userRegister, singIn, sendVerifyUserEmail} from '../src/services/authServices.js';
+import {singup, singIn} from '../src/services/authServices.js';
+import {sendVerifyUserEmail} from '../src/utilities/emailDelivery.js';
 import {validationResult, check} from 'express-validator';
 import {ResponseDto} from '../src/dtos/responseDto.js';
 import jwt from 'jsonwebtoken';
@@ -50,7 +51,7 @@ router.post('/user/sendVerifyEmail', emailValidations, async (req, res) => {
   res.send(resultProcess);
 });
 
-router.post('/user/userregister', RegisterValidationRules, async (req, res)=>{
+router.post('/user/singup', RegisterValidationRules, async (req, res)=>{
   const errors = validationResult(req);
   if (!errors.isEmpty()) {
     const response = new ResponseDto(400, null, errors);
@@ -58,8 +59,8 @@ router.post('/user/userregister', RegisterValidationRules, async (req, res)=>{
   }
   const requestData = new ReqSignUpDto(req.body.username, req.body.email,
       req.body.fullname, req.body.password);
-  const userRegisterResult = await userRegister(requestData);
-  res.send(userRegisterResult);
+  const singupResult = await singup(requestData);
+  res.send(singupResult);
 });
