+++ server/src/services/guessNumberServices.js
@@ -32,7 +32,7 @@ async function checkAnswer(guess) {
     result.status = 1;
     result.randomNumber = randomNumber;
     const score = calculateScore(chance);
-    const saveRes = await SaveRecord(new PlayDto(1, gameId, score));
+    const saveRes = await saveRecord(new PlayDto(1, gameId, score));
     if (saveRes === true) {
       return result;
     } else {
@@ -51,7 +51,7 @@ async function checkAnswer(guess) {
   return result;
 }
 
-async function SaveRecord(playDto) {
+async function saveRecord(playDto) {
   try {
     const play = await prisma.play.create({
       data: {
@@ -82,6 +82,6 @@ async function getGameIdByName(gameName) {
   } catch (error) {
     console.log(error.message);
   }
-  return null;
+  return undefined;
 }
 export {checkAnswer, restartGame};
